+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### VNEXT
 
+* Allow passing in a string `schema` to `makeRemoteExecutableSchema` [PR #521](https://github.com/apollographql/graphql-tools/pull/521)
 * ...
 
 ### v2.11.0